requirements.txt
clip==1.0
numpy==1.24.1
Pillow==9.4.0
Pillow==10.3.0
scikit_learn==1.3.0
torch==2.2.0
torchvision==0.15.1
tqdm==4.65.0
transformers==4.37.2
